backend/requirements.txt
pantheon
autopep8
aiofiles
compress-json
fastapi
httpx
ipykernel
ipywidgets
jinja2
jupyterlab
kmodes
matplotlib
numpy
notebook
pandas
Pillow
pydantic
python-dotenv
pymongo[srv]
motor
python-multipart
pytest
pytest-asyncio
requests
riotwatcher
seaborn
# sentence-transformers
scikit-learn
scikit-optimize
scikeras
scipy
threadpoolctl
uvicorn
xgboost